Add OnUninitialize lifecycle method
This implements support for a new lifecycle method on AceAddon objects
named 'OnUninitialize' that will be called whenever ADDONS_UNLOADING
fires, and will be dispatched in reverse order of initialization.

The use case here is for addons mainly aimed at addons that maintain
separation between their runtime and saved state, and want to defer
writing this state to saved variables until it's actually needed.

While at a high level an addon can just trivally register for this
event and handle it themselves, things get a bit more awkward when it
comes to ordering - event dispatch is unordered, so an addon with
multiple submodules would have de-initialization logic run in an
effectively random order. Implementing this lifecycle method natively
within AceAddon allows a sane ordering to be trivially defined.