Upgrade Ring metatables on library upgrades
This is hopefully a non-risky and better change going forward, but don't
want to rock the boat immediately after fixing one upgrade issue.

As changing the implementation of any Ring method in a current version
means it has no effect on rings created in older versions, it makes
sense to apply the new metatable to avoid this trap in the future.

This appears to work fine from an upgrade of version 24 to 28 (which
adds the Link method to Ring).