Release message data for GCing
Believe this code was just never updated to accomodate 4.x changes
that added prefixes to addon messages; currently for those it's
nilling out the prefix field and not the message text.

This isn't strictly *required* as deleted messages sit in a weak
bin and as such the strings should be considered unreachable, but
as the existing code was doing a nil assignment anyway it makes
sense to fix this up.